Crate solana_metrics

Re-exports

pub use crate::metrics::flush;

pub use crate::metrics::query;

pub use crate::metrics::set_host_id;

pub use crate::metrics::set_panic_hook;

pub use crate::metrics::submit;

Modules

counter

datapoint

This file defines a set of macros for reporting metrics.

metrics

The metrics module enables sending measurements to an InfluxDB instance

Macros

create_counter

create_datapoint

datapoint

datapoint_debug

datapoint_error

datapoint_info

datapoint_trace

datapoint_warn

inc_counter

inc_counter_info

inc_new_counter

inc_new_counter_debug

inc_new_counter_error

inc_new_counter_info

inc_new_counter_warn

Structs

CounterToken

A token for TokenCounter.

MovingStat

TokenCounter

A helper that sends the count of created tokens as a datapoint.