solana_metrics

Module metrics

The metrics module enables sending measurements to an InfluxDB instance

Modules

test_mocks

Structs

MetricsAgent

Enums

MetricsError

Traits

MetricsWriter

Functions

flush

Blocks until all pending points from previous calls to submit have been transmitted.

get_host_id

metrics_config_sanity_check

query

serialize_points

set_host_id

set_panic_hook

Hook the panic handler to generate a data point on each panic

submit

Submits a new point from any thread. Note that points are internally queued and transmitted periodically in batches.